api/requirements.txt
django
django-cors-headers
django-filter
djangorestframework
psycopg2
python-dotenv
requests